derive: use fewer unwraps
`rinja_derive(_standalone)` expects that the environment variable
`CARGO_MANIFEST_DIR` is set and panics otherwise. It also expects that
the input can be parsed as a struct/enum/union and panics otherwise.

This PR uses `"."` as fallback path if `CARGO_MANIFEST_DIR` is unset,
and generates a compiler error if the input TokenStream could not be
parsed.

Both error conditions should only occur in `rinja_derive_standalone`,
esp. if you have an unexpected runtime like WASM that does not have a
filesystem or environment variables.